Add locations from 2013
Added a list of places and their locations from a map
which was present on the website till the year 2013.
This list was obtained from a source code of the page with map found at
Wayback Machine (https://web.archive.org/web/20130417101525/http://www.kafelanka.cz:80/mista/mapa.php).
The list is stored as a CSV file.